requirements.txt
Flask==2.0.3
torch==1.9.0
Pillow==8.2.0
ultralytics==8.0.0
pandas==1.3.3
gunicorn==20.0.4
jinja2==2.11.3
werkzeug==1.0.1
marksafe==1.1.1